Rename replication's seq_ to next_seq for clarity
and remove slave's replication progress info in INFO cmd, as the self->seq_
is not updated for every received batch and the progress info is been
presented in master's INFO cmd already. Also the slave's replication
is the same as the slave's db's sequence, so no need to duplicate that
info.