Don't use extra port for replication (#200)
As we know, kvrocks has specified replication threads that use 'port + 1' as
listening port and process replication requests, that may make somethings easy,
but it is not convenient to deploy and maintain. In this commit, kvrocks doesn't use
extra port for replication but process replication requests in general work threads.

In order to be compatible with the old version, we provide a configuration item
'master-use-repl-port' to make new version be able to replicate old version master.
If the masters are old version, you should set 'master-use-repl-port' to yes,
otherwise you should set it to no.