Add config purge-backup-on-fullsync
If replicas need full synchronization with maste, master need to create
checkpoint for feeding replicas, and replicas also stage a checkpoint of
the maste. If we also keep the backup, it maybe occupy extra disk space.
You can enable 'purge-backup-on-fullsync' if disk is not sufficient, but
that may cause remote backup copy failing. This config is no by default.

Other changes
  - Simply how to check checkpoint exists or not
  - Adjust cron for purging backup and checkpoint