Adapt redis sentinel ping check (#316)
In redis sentinel, it sends ping command to the instance and check response,
kvrocks in replica role will return an error that is not LOADING when restoring
dataset from backup, so sentinel doesn't update the replica instance available
time and set the instance subjectively down. But for master, there is no problem.