Remove exposing slot-id-encoded config (#330)
Before, we expose slot-id-encoded in config file, that is useful to migrate
keys based on slot when cluster is enabled, actually, we should only enable
slot-id-encoded when we want to use cluster mode. So exposing this config may
make users confused, and wrong config also make data corrupt. Now we remove
this config, and set it automatically.