Support GET command on bitmap type (#447)
In Kvrocks, bitmap and string are two completely different types. Although setbit,
getbit, bitcount, and bitpos operations on the string type are supported.
@smartlee once mentioned that when using setbit on the string type, master-slave
replication will have a tremendous amount of write amplification because Kvrocks
replication does not use command propagation like Redis. In addition, using bitmap
type conserves storage space. So we recommend using type bitmap.
But to make it easy to use, we want to get the whole string after setbit operations,
so we implement this interface.

We covert bitmap to string when executing GET command and return string reply,
otherwise, we provide one configuration item max-bitmap-to-string-mb which uses
to limit the max size of bitmap to string transformation(MB).

Why we need to swap the bit when converting bitmap type to string?
If you setbit bit 0 1, the value is stored as 0x01 in Kvocks but 0x80 in Redis.
So swapping bits is to get the same return value as Redis.