Avoid flush all redis db in kvrocks2redis (#498)
In the kvrocks2redis tool, when kvrocks sends data to Redis, it will flush
the whole target redis by using the FLUSHALL command. We shouldn't do that,
we just need to clear the target DB.

In addition, modify the comments in the configuration file to make it more
accessible. And based on the output of cpplint, fixed the nonstandard header file.