Use FetchContent instead of ExternalProject and git submodules in CMake (#564)
Currently, we are using git submodules to manage external libraries which
need to pull full git histories. CMake gives a light way `FetchContent` to handle
this scenario after 3.11, we can only fetch the specify tag or commit  instead of full histories. 

Another minor change was that the CMake minimal required will be 3.13 to support the 
FetchContent instruction.

Co-authored-by: tison <wander4096@gmail.com>
Co-authored-by: hulk <hulk.website@gmail.com>