Fix RocksDB can't auto resume after disk quota exceeded error (#628)
In #229, the issue where RocksDB could not recover from the no Space background
error was fixed. This problem RocksDB at facebook/rocksdb#8376 has been repaired,
but the issue has not been thoroughly solved, The same problem will still occur
when an EDQUOT Disk Quota Exceeded error is encountered (see the detailed in 
facebook/rocksdb#10134).

RocksDB cannot recover from this problem and must be restarted. This problem is
more likely to occur when kvrocks is deployed in container.

In order to handle all versions of RocksDB, we manually resume DB when we encounter
two retryable io errors: No space left on device and Disk Quota Exceeded.

For the Disk Quota Exceeded error, RocksDB did not expose a friendly interface,
so we did a string match.