Allow to enable the async_io option to improve the performance (#1215)
Currently, RocksDB supports the async io to mitigate the impact of storage latency
to read asynchronously and in parallel as much as possible to hide IO latency. And it
only takes effect in MultiGet and Iterators:

In Iterators, it will prefetch data asynchronously in the background for each file being iterated on. 
In MultiGet, it will read the necessary data blocks from those files in parallel as much as possible.

For the current implementation, we only allow enabling the async io in Iterator, because the MultiGet
operation depends on the folly library and C++ 20 which it's NOT expected to introduce in Kvrocks.
---------

Co-authored-by: hulk <hulk.website@gmail.com>
Co-authored-by: Twice <twice.mliu@gmail.com>