Add an optional args to XLEN command to specify starting ID and direction
Syntax: `XLEN stream_name [entry_id] [direction]`.
Where direction can be '-' or '+' (special IDs mean respectively the first and the last elements in the stream). Default value: '+'.
If an optional entry ID is specified, the XLEN command returns the number of entries between that ID and the first/last element of the stream.
The element with the specified ID is not considered - it serves as an exclusive boundary.
If the stream doesn't contain an entry with the specified ID, the command starts counting from the closest entry to the first/last element in the stream.
If the specified ID is equal to or greater than the last entry ID and:
a) direction is '+' - the command returns 0
b) direction is '-' - the command returns stream's size
If the specified ID is equal to or less than the first entry ID and:
a) direction is '-' the command returns 0
b) direction is '+' the command returns stream's size
This feature is helpful in finding:
a) how many entries in the stream were processed so far
a) how many unprocessed (ahead of the current one) entries are in the stream
Complexity: O(n)