Add SUBSTR redis command
This PR adds the SUBSTR command, command syntax is exactly
the same as GETRANGE. Although SUBSTR is already deprecated
(It can be replaced by GETRANGE), we still want to support it.

Note that go-redis does not support the SUBSTR command, so rdb.Do
is used in go tests to test it.