Add a particular test case for HGETALL with async_io=true
Was reproducible with RocksDB v8.0.0, large values in a hash set (>512 bytes).
After the COMPACT command was executed, HGETALL/HVALS/HKEYS couldn't "see"
all the elements of the hash set. However, HGET could find those values.
It looked like the iterator via the snapshot didn't see them for some reason.
Since the version of the RocksDB was bumped to v8.0.1, the bug disappeared.