Check if the cluster migrate sync command will get error
Currently, the migration timeout test case failed frequently when
running in GitHub Actions and the value result is Nil. But we cannot
know if it's caused by error or the result value is a nil string, so we
can need to check if it occurs error to narrow the scope.