Change the migration timeout type from float to int (#1464)
Currently, the migration timeout is a float and allows using 1.2 to represent
the 1.2s, but it's a bit inconsistent with other timeouts like BRPOP which requires
the timeout to be an integer. So to make the timeout more consistent, we would like
to change the migration timeout from the float to int as well.

This PR also fixed the flaky test since the previous migration timeout 0.5s was too short
and cause test case failure frequently.