Make INFO command reject cases with arguments greater than 2 (#1570)
Passing info a b c is the same as passing plain info
since we only checked that args_.size() == 2, by default it
will output all.

In this PR, we will reject args_.size() > 2 and return a
syntax error. This is consistent with Redis 6.2 behavior,
however, it is inconsistent with Redis 7.0, Redis 7.0 supports
passing info server clients...

Because we don't plan to do such support for the time being,
we will first fix the problematic parts.