Fix flaky test TestZset ZRANGEBYSCORE_fuzzy_test
The old code used fmt.Sprintf("%f") to handle float,
it will have a loss of precision.

A simple go example:
```
func main(){
	min1, max1 := rand.Float64(), rand.Float64()
	fmt.Println(min1) // 0.6046602879796196
	fmt.Println(max1) // 0.9405090880450124

	min2, max2 := fmt.Sprintf("%v", min1), fmt.Sprintf("%v", max1)
	fmt.Println(min2) // 0.6046602879796196
	fmt.Println(max2) // 0.9405090880450124

	min3, max3 := fmt.Sprintf("%f", min1), fmt.Sprintf("%f", max1)
	fmt.Println(min3) // 0.604660
	fmt.Println(max3) // 0.940509
}
```

As we can see in the example, the min / max is inaccurate
after the %f handle. So in our test, it is ok get the unexpected
result in range.

i.e. min, max are actually accurate, but the arguments we
use for zrange is actually inaccurate.

Fix #1552