Remove deprecated golangci linters: structcheck, deadcode and varcheck
Ci resulted in the following warnings:
```
level=warning msg="[runner] The linter 'structcheck' is deprecated
(since v1.49.0) due to: The owner seems to have abandoned the linter.
Replaced by unused."
level=warning msg="[runner] The linter 'deadcode' is deprecated (since
v1.49.0) due to: The owner seems to have abandoned the linter.  Replaced
by unused."
level=warning msg="[runner] The linter 'varcheck' is deprecated (since
v1.49.0) due to: The owner seems to have abandoned the linter.  Replaced
by unused."
```

The deadcode, structcheck and varcheck linters are deprecated by
golangci-lint. It is recommended to use unused instead. And we are
already using it, so this PR just removes the deprecated linters.