Fix data race when updating the slots_info string
Currently, we cached the slots_info inside the cluster node
and update the slots_info every time before getting the node
description. But we didn't use the execlusive lock except the cluster
set command, so it may have the data race between the read and write
operation and cause crash. This fix avoid this by removing the
unnecessary cache field `slots_info` from the cluster node.