Enable plain merge in enabled_merge_buttons
In some cases, such as a release version, we will pick some
commits, then do a PR review, and then merge the commits into
the branch. In this case, we want to use plain merge to keep
the info of each commit instead of using squash-merge.

Or in some refactoring PRs, we can separate code moves (rename)
and actual changes into different commits. These single commits
we want to keep the commit message and doing a plain merge.
In this way, when tracking changes, we can only focus on actual
changes without being affected by meaningless diffs.

When enabled, we can control it (when we are doing merge, usually
squash is used, and in some cases we need to keep commits we use merge).
That is, the choice of squash or merge is the responsibility of the
person for merging the PR.

See #1622 for more discussion on release part.