Fix shouldn't try to create column familes when opening with the readonly mode
Currently, kvrocks2redis will read and parse data from local DB with
openning in readonly mode, and it will be failed if the Kvrocks is
running on the same DB dir. The root cause is the running Kvrocks would
acquire the DB lock and kvrocks2redis would try to acquire this DB
lock as well when creating column familes.