Fix RESTORE TTL is 0 with ASBTTL does not restore the key (#1724)
TTL is 0 mean the key is created without any expire.
ABSTTL mean the TTL should represent absolute Unix timestamp
in milliseconds in which the key will expire.

When typing TTL 0 and ABSTTL, kvrocks will not restore the
key, since we think the key is expired. However, in Redis,
ABSTTL is ignored when TTL is 0. (Although they seem to be
mutually exclusive)