Introduce MDel do the batch delete and use MultiLockGuard guarantee atomicity (#1712)
Previously, the DEL/UNLINK commands looped through each key to
delete a single key. We need to change to batch delete (write),
and also the original implementation did not lock the keys and
will break atomicity.

This PR introduces a new Mdel function, which uses MultiLockGuard
to lock all the keys to guarantee atomicity, and use batch to do
the final delete (write).

Now it will be only used in DEL / UNLINKE commands. Noticed that
our DEL / UNLINK test coverage was not sufficient, added some simple
coverage tests to cover it.

Fixes for DEL / UNLINK issues in #1692.

unrelated change: in Exists for loop Get, return an execution error
when `!s.ok() && !s.IsNotFound()` without continuing the operation
(because it will be some internal RocksDB error).