Don't rewrite unchanged items into the configuration file
Currently, config rewrite will also add some default key-values
like backup-dir, pidfile into the configuration file. Some of them
are for compatiable purpose and don't expected to be seen by users.

So we resolve this by ignoring the configuration items if there'are
not existing in the configuration file and the value was not modified.