Don't rewrite unchanged items into the configuration file (#1882)
Currently, config rewrite will also add some default configurations 
like backup-dir, pidfile into the configuration file. Some of them
are for compatible purposes and don't expect to be seen by users.

So we resolve this by ignoring the configuration items
if they do not exist in the configuration file and the value was not modified.