Graceful shutdown the workers when reducing worker threads (#1863)
#1855 introduced the data race even only migrating the non-blocking connections. For example:

T1: C0(connection) is running the command Config::Set on W0(worker)
and C1 is running another command on W1. C0 got the exclusive lock for executing commands
and C1 will wait for the lock inside ExecuteCommands.

T2: C0 migrates C1 to W0 after reducing the number of worker threads, but the ExecuteCommands function will
continue executing after migrating. So both W0 and W1 will access the C1 at the same time.

To avoid this, I simply don't migrate the connection if it has any running command and
delay to shutdown workers, so that old connections have a chance to finish the running command.