Add redis_mode/kvrocks_mode fields in INFO SERVER (#1927)
Although we will already show cluster_enabled in INFO CLUSTER,
some tools such as Redis Manager will need this field to determine
whether it is cluster mode.

If cluster mode is enabled, redis_mode is displayed as a cluster,
otherwise, it is standalone. We also use this opportunity to add
a kvrocks_mode field.

Fixes #1926.