fix(config): avoid rewriting the config file if it's unnecessary (#2347)
The server will start by rewriting the config file to persist namespace/token pairs
if namespace replication is enabled. But it's unnecessary if the namespace replication
is disabled or no tokens are loaded from the configuration file because the purpose of
this rewrite is to remove tokens from the config file and write them to the database.