Fix didn't reload namespaces from DB after the full sync
Namespaces will store inside DB if the repl-namespace-enabled was set to
yes. And then replicas will intercept and parse the increment replication log
to see if it needs to reload namespaces. But it won't have the namespace
update log when doing the full sync, so we need to forcely reload from
DB once the full replication was done.