fix(replication): didn't resume the db status after restarting full sync (#2549)
Currently, the pre_fullsync_cb will stop the task runner and set the DB loading status to yes,
but it didn't resume those states. This will cause the server to run in restoring status until
success in resyncing from the master. To fix this, we need to call the post_fullsync_cb to resume those statuses
before restarting full sync.

This PR also uses try_lock to allow the replication thread to be stopped while preparing the restore db.