Fix potential data race between blocking command and transaction
Currently, the EXEC is an exclusive command and will enter the
transaction mode when running, but the blocking command's callback
function doesn't have this lock. So the callback might trap into wrong
transcation mode and broken the write batch.

For example, the client A sends the command: `BRPOP list 10` and it will
enter the blocking mode due to no data. And the client B use the
transaction to send the command: `MULTI; LPUSH list a; EXEC`. Then the
client A's callback will be revoked after the `LPUSH` command was
executed but the EXEC is running. So the callback's operation like
removing an element from list will be counted into transaciton.